Fix the bitmap issue when there is a gap in serialization ID of the proto
The current implemention assumes serialization ID is continuous. When
there is gap between serialization ID, there will be runtime error of `xxx class
doesn't have attribute bitmap<Xi> attribute` when we try to create

This is because the bitmap is computed by the length of the fields. But
the position is computed by the serialization ID value. We should use
the max serialization ID to compute the number of bitmap.